regression/requirements.txt
Flask==3.0.3
Flask-CORS== 5.0.0
nltk==3.9.1 
numpy==2.2.5
pandas==2.2.3
psycopg2-binary==2.9.10
python-dotenv==1.1.0
matplotlib==3.10.3
seaborn==0.13.2
scikit-learn==1.6.1
SQLAlchemy==2.0.41
statmodels==0.14.4
